Move sample_uniform of MatZ to submodule
